Refactor __version__.py into __init__.py
Cleans up global variable location and aligns to common practice.
Additionally cleans up remnants from py27 deprecation.
