--- tokendito/__version__.py
@@ -1,11 +1,0 @@
-# vim: set filetype=python ts=4 sw=4
-# -*- coding: utf-8 -*-
-"""tokendito version."""
-__version__ = "2.0.0"
-__title__ = "tokendito"
-__description__ = "Get AWS STS tokens from Okta SSO"
-__long_description_content_type__ = "text/x-rst"
-__url__ = "https://github.com/dowjones/tokendito"
-__author__ = "tokendito"
-__author_email__ = "tokendito@dowjones.com"
-__license__ = "Apache 2.0"
