+++ tox.ini
@@ -44,8 +44,3 @@ format = ${cyan}%(path)s${reset}:${yellow_bold}%(row)d${reset}:${green_bold}%(co
 [pytest]
 env =
     PIP_DISABLE_PIP_VERSION_CHECK=1
-    PYTHONWARNINGS=ignore:DEPRECATION, ignore::UserWarning, ignore::DeprecationWarning
-filterwarnings =
-    ignore::DeprecationWarning
-    ignore::UserWarning
-    ignore:DEPRECATION
